services/ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.9
python-dotenv==1.0.0
pydantic==2.13.0
pydantic-settings==2.13.1

# LLM & Embeddings
litellm==1.3.1
langchain==1.2.15
sentence-transformers==5.4.1

# Vector Database
chromadb==1.5.7

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11

# Database
pymongo==4.6.0
motor==3.3.2

# HTTP & Utilities
httpx>=0.27.0,<0.29.0
aiohttp==3.9.1

# Logging & Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Code Quality
pylint==3.0.3
black==23.12.0

# Service Registry
python-consul2==0.1.5
